pyproject.toml
[tool.poetry]
name = "kamma"
version = "0.1.0"
description = ""
authors = ["erfanhresh <erfan_bcln@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.1"
djangorestframework = "^3.16.0"
django-autoutils = "^2.1.1"
django-colorfield = "^0.14.0"
python-decouple = "^3.8"
django-debug-toolbar = "^5.2.0"
psycopg2-binary = "^2.9.10"
drf-yasg = "^1.21.10"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^8.1.0"}
django-filter = "^25.1"
drf-nested-routers = "^0.94.2"
django-constance = "^4.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
